% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACPcalculator.R
\name{ACP.calculator}
\alias{ACP.calculator}
\title{Antecedental Accumulation calculator for the VI Complete Time Series}
\usage{
ACP.calculator(CTSR.VI, ACP.table)
}
\arguments{
\item{CTSR.VI}{Complete Time Series of Vegetation Index. An object of class \code{'ts'}. Monthly time series of VI values}

\item{ACP.table}{A table of every combination of offset period and accumulation period. if ACP.table = FALSE, CTSR.RF and acu.RF must be provided}
}
\value{
summary
(To be filled in)

CTSR.precip (CTSR.RF)
Complete Time Series of the optimally accumulated rainfall.An object of class 'ts'.
}
\description{
Takes the Complete Time Series Vegetation Index and a table of every possible accumulation period and offset
period.  A OLS is calculated \code{\link{lm}} for every combination of VI ~ rainfall.  This Function preferences those results where
slope>0 (increase in rainfall causes an increase in vegetation), returning the rainfall accumulation that has the highest R-squared
and a positive slope. If no combinations produce a positive slope then the one with the highest Rsquared is returned.
}

