% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_pos.R
\name{AnMax.VI}
\alias{AnMax.VI}
\title{Annual max VI Calculator}
\usage{
AnMax.VI(CTSR.VI)
}
\arguments{
\item{CTSR.VI}{Complete Time Series of Vegetation Index. An object of class \code{'ts'}. Monthly time series of VI values}
}
\value{
Max(anu.VI)
The annual (Growing season) max VI.

Max.Month
a Vector that has the month number where max values wew observed

index(VI.index)
the index of the CTSR.VI ts that the anu.VI values occur at.R indexs from 1 rather than 0.
}
\description{
Takes the montly time series of the VI and calculates the growing season max VI. In series where the peak occurs in November or December,
an interannual growing season is assessed.
}

