% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annual_precipitation.R
\name{AnnualRF.Cal}
\alias{AnnualRF.Cal}
\title{Antecedental Accumulation calculator for the Annual Max VI Time Series}
\usage{
AnnualRF.Cal(anu.VI, VI.index, ACP.table, Breakpoint = FALSE,
  allow.negative = FALSE)
}
\arguments{
\item{anu.VI}{The annual (Growing season) max VI. Which can be calculated from the CTSR.VI using \code{\link{AnMax.VI}}.}

\item{VI.index}{the index of the CTSR.VI ts that the anu.VI values occur at. NOTE. R indexs from 1 rather than 0.
NOTE. if VI.index=FALSE, It can be calculated from the CTSR.VI using \code{\link{AnMax.VI}}.}

\item{ACP.table}{A table of every combination of offset period and accumulation period. It can be calculated from ________}
}
\value{
_____(acu.RF)
The optimal accumulated rainfall for anu.VI. Mut be a object of class 'ts' and of equal length to anu.VI.
if anu.RF=FALSE, it will be calculated from ACP.table. see(____)

summary
(To be filled in)
}
\description{
Takes the Annual Max VI Time Series, the VI.index and a table of every possible accumulation period and offset
period.  A OLS is calculated \code{\link{lm}} for every combination of VI ~ rainfall.  This Function preferences those results where
slope>0 (increase in rainfall causes an increase in vegetation), returning the rainfall accumulation that has the highest R-squared
and a positive slope. If no combinations produce a positive slope then the one with the highest Rsquared is returned.MISSING non
peramtric and other variables
}

