% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TSSRESTREND.R
\name{plot.TSSRESTREND}
\alias{plot.TSSRESTREND}
\title{Plot function for class TSSRESTREND class}
\usage{
\method{plot}{TSSRESTREND}(x, verbose = FALSE, plots = "all", ...)
}
\arguments{
\item{plots}{= "all"
User can create a single plot,}
}
\description{
Plot *******
}

