% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TSSRESTREND.R
\name{print.TSSRESTREND}
\alias{print.TSSRESTREND}
\title{Print function for class TSSRESTREND class}
\usage{
\method{print}{TSSRESTREND}(x, ...)
}
\arguments{
\item{x}{Object of class TSSRESTREND}

\item{...}{further arguments passed to the function.}
}
\description{
print function for class TSSRESTREND
}

