% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbitACPtable}
\alias{rabbitACPtable}
\title{Rabbit Impacted Vegetation Precipitation Accumulation Table}
\format{
A matrix containing the complete time series of every acp and ofp. See \code{\link{climate.accumulator}} for details
}
\source{
Awap data from \url{http://www.csiro.au/awap/cgi/awap2.pl}
}
\usage{
rabbitACPtable
}
\description{
Rainfall Accumulation Table form a part of the Simpson-sStrzelecki Dunefields bioregion
impacted by reduced rabbit predation after the release of RHD.
}
\examples{
data(rabbitACPtable)
}
\keyword{datasets}
