% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTREND.R
\name{RESTREND}
\alias{RESTREND}
\title{RESTREND (RESidual TREND)}
\usage{
RESTREND(anu.VI, acu.RF, VI.index, sig = 0.05, print = FALSE,
  plot = FALSE)
}
\arguments{
\item{anu.VI}{The annual (Growing season) max VI. if anu.VI=FALSE, it will be calculated from the CTSR.VI. See (___)}

\item{acu.RF}{The optimal accumulated rainfall for anu.VI. Mut be a object of class 'ts' and of equal length to anu.VI. if anu.RF=FALSE, it will be calculated from ACP.table. see(____)}

\item{VI.index}{the index of the CTSR.VI ts that the anu.VI values occur at. Must be the same length as anu.VI. NOTE. R indexs from 1 rather than 0.}

\item{sig}{Significance of all the functions, sig=0.05}

\item{print}{Prints more details at every step of the procces}

\item{plot}{creates a plots of the residulas vs time}
}
\value{
summary
(To be filled in)

VPR
the lm (See lm_______)
}
\description{
For ts with no significant breakpoints in the residuals or the VPR. Takes annual VI max and it associated optimal accumulated precipitation
}
\author{
Arden Burrell, arden.burrell@unsw.edu.au
}

