% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segVPR.R
\name{seg.VPR}
\alias{seg.VPR}
\title{Segmented Vegetation Precipitation Relationship}
\usage{
seg.VPR(anu.VI, acu.RF, VI.index, breakpoint, rf.b4, rf.af, sig = 0.05)
}
\arguments{
\item{anu.VI}{The annual (Growing season) max VI. if anu.VI=FALSE, it will be calculated from the CTSR.VI. See (___)}

\item{acu.RF}{The optimal accumulated rainfall for anu.VI. Mut be a object of class 'ts' and of equal length to anu.VI. if anu.RF=FALSE, it will be calculated from ACP.table. see(____)}

\item{VI.index}{the index of the CTSR.VI ts that the anu.VI values occur at. Must be the same length as anu.VI. NOTE. R indexs from 1 rather than 0.}

\item{breakpoint}{TO BE ADDED}

\item{sig}{Significance of all the functions, sig=0.05}

\item{print}{Prints more details at every step of the procces}

\item{plot}{creates a plots of the VPR vs rainfall standard variance}
}
\value{
summary
(To be filled in)

VPR
the lm (See lm_______)

TSS.RESTREND
FALSE
}
\description{
For ts with a significant breakpoints in the the VPR. Takes annual VI max, the optimal accumulated precipitation
bor and after the breakpoint, then caculated the Precipitation Standard Variance, and regresses it with a dummy
variable that is 0 before the breakpoint and 1 after it. WARNING NEED TO REMOVE acu.RF
}
\author{
Arden Burrell, arden.burrell@unsw.edu.au
}

