% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segRESTREND}
\alias{segRESTREND}
\title{Data frame containing the annual data for a segRESTREND analysis}
\format{R data frame}
\source{
\code{\link[gimms]{gimms-package}}
}
\usage{
segRESTREND
}
\description{
contains anu.VI, acu.RF,  VI.index, rf.b4, rf.af. Range 1982 - 2013. Breakpoint for this pixel is 24 (2005)
}
\keyword{datasets}
