% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segRESTRENDCTSR}
\alias{segRESTRENDCTSR}
\title{Data frame containing the Complete Times Series data for a segmented RESTREND analysis}
\format{R data frame}
\source{
a single pixel from \code{\link[gimms]{monthlyComposite}}
}
\usage{
segRESTRENDCTSR
}
\description{
contains CTSR.VI and CTSR.RF. See \code{\link{TSSRESTREND}}. Starts 1/1982 and extend to 12/2013
}
\seealso{
\code{\link{stdRESTREND}} for the annual values for the same pixel
}
\keyword{datasets}
