% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segRESTRENDrfTab}
\alias{segRESTRENDrfTab}
\title{Precipitation Accumulation Table for the segRESTREND demonstration pixel}
\format{
A matrix containing the complete time series of every acp and ofp. See \code{\link{climate.accumulator}} for details
}
\source{
Awap data from \url{http://www.csiro.au/awap/cgi/awap2.pl}
}
\usage{
segRESTRENDrfTab
}
\description{
Rainfall Accumulation Table
}
\keyword{datasets}
