% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segVPRctRF}
\alias{segVPRctRF}
\title{Data frame containing the raw rainfall data set for the segVPR data,  ending dec 2013 with a frquency of 12}
\format{R data frame}
\source{
Awap data from \url{http://www.csiro.au/awap/cgi/awap2.pl}
}
\usage{
segVPRctRF
}
\description{
contains raw monthly precipitation
}
\seealso{
\code{\link{stdRESTREND}} for the annual values for the same pixel
}
\keyword{datasets}
