% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stdRESTREND}
\alias{stdRESTREND}
\title{Data frame containing the annual data for a standard Restrend analysis}
\format{R data frame}
\source{
\code{\link[gimms]{gimms-package}}
}
\usage{
stdRESTREND
}
\description{
contains anu.VI, acu.RF and VI.index. Range 1982 - 2013
}
\keyword{datasets}

