% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stdRESTRENDctRF}
\alias{stdRESTRENDctRF}
\title{Data frame containing the raw rainfall data set ending dec 2013 with a frquency of 12}
\format{
R data frame
}
\usage{
stdRESTRENDctRF
}
\description{
contains raw monthly precipitation
}
\seealso{
\code{\link{stdRESTREND}} for the annual values for the same pixel
}
\keyword{datasets}
