% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stdRESTRENDrfTab}
\alias{stdRESTRENDrfTab}
\title{Precipitation Accumulation Table for the standard RESTREND demonstration pixel}
\format{A matrix containing the complete time series of every acp and ofp. See \code{\link{rainfall.accumulator}} for details}
\source{
Awap data from \url{http://www.csiro.au/awap/cgi/awap2.pl}
}
\usage{
stdRESTRENDrfTab
}
\description{
Rainfall Accumulation Table
}
\keyword{datasets}

