% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.RESTRENDpackage.R
\docType{package}
\name{TSS.RESTREND-package}
\alias{TSS.RESTREND}
\alias{TSS.RESTREND-package}
\title{TSS.RESTREND: Time Series Segmented RESidual TREND}
\description{
The TSS-RESTREND package is designed used via the TSS-RESTREND function which calls all the other functions in the
correct order to perform a complete TSS-RESTREND analysis.  The individual functions have been made callable to
allow for greater user control.  The version in active development, as well as additional documentation, data and
example scripts, see \url{https://github.com/ArdenB/TSSRESTREND}.
}
\details{
Version 0.3.0
}
\note{
This code is a demostration of the method and not the code used in Burrell et. al., (2017).
       The original code uses both python and R and was designed for batch proccessing. For the pixels
       tested using the TSS.RESTREND package it produces identical results.
}
\author{
\strong{Maintainer}: Arden Burrell \email{arden.burrell@unsw.edu.au}

}
