% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS_RESTREND.R
\name{TSS.RESTREND}
\alias{TSS.RESTREND}
\title{Time Series Segmentation of Residual Trends}
\usage{
TSS.RESTREND(CTSR.VI, ACP.table = FALSE, CTSR.RF = FALSE, anu.VI = FALSE,
  acu.RF = FALSE, VI.index = FALSE, rf.b4 = FALSE, rf.af = FALSE,
  sig = 0.05, print = FALSE, plot = TRUE, details = FALSE)
}
\description{
Takes in a complete monthly time series of a VI and its corrosponding precipitation
}

