% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viMAX_plot.R
\name{VImax.plot}
\alias{VImax.plot}
\title{Annual VI Max ploter}
\usage{
VImax.plot(anu.VI, brkp = FALSE)
}
\arguments{
\item{anu.VI}{The annual (Growing season) max VI. if anu.VI=FALSE, it will be calculated from the CTSR.VI. See (___)}

\item{brkp}{A detected significant breakpoint in the VPR or in the VPR-residuals.}
}
\value{
plot
}
\description{
Plots the VI max time series
}

