% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.RESTRENDpackage.R
\docType{package}
\name{TSS.RESTREND-package}
\alias{TSS.RESTREND}
\alias{TSS.RESTREND-package}
\title{TSS.RESTREND: Time Series Segmented RESidual TREND}
\description{
TSS.RESTREND This script is a a demonstration of the Segmented RESTREND method proposed in Burrell et. al., (2016?).
(Paper submitted). TSS.RESTREND is a modification of the RESTREND method originally proposed by Evans and Geerke (2006).
It incoperates BFAST breakpoint detection to look for major ecosystem changes.  The version in development and additional
documentation, data and example scripts can be can be found at \url{https://github.com/ArdenB/TSSRESTREND}.
}
\details{
Version 0.2.00
}
\note{
This code is a demostration of the method and not the code used in Burrell et. al., (2016?).
       The original code uses both python and R and was designed for batch proccessing. For the pixels
       tested using the TSS.RESTREND package it produces identical results.
}

